+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/Axis/CustomizingTicks.cs
@@ -168,7 +168,8 @@ public class CustomizingTicks : ICategory
             myPlot.Add.Signal(Generate.Sin());
             myPlot.Add.Signal(Generate.Cos());
 
-            myPlot.Grid.MajorLineStyle.Width = 1; // TODO: demonstrate how to disable just vertical or horizontal grid lines
+            myPlot.Grid.XAxisStyle.IsVisible = true;
+            myPlot.Grid.YAxisStyle.IsVisible = false;
         }
     }
 
@@ -241,9 +242,9 @@ public class CustomizingTicks : ICategory
             myPlot.Axes.Left.TickGenerator = tickGen;
 
             // show grid lines for minor ticks
-            myPlot.Grid.MajorLineStyle.Color = Colors.Black.WithOpacity(.15);
-            myPlot.Grid.MinorLineStyle.Color = Colors.Black.WithOpacity(.05);
-            myPlot.Grid.MinorLineStyle.Width = 1;
+            myPlot.Grid.MajorLineColor = Colors.Black.WithOpacity(.15);
+            myPlot.Grid.MinorLineColor = Colors.Black.WithOpacity(.05);
+            myPlot.Grid.MinorLineWidth = 1;
         }
     }
 }