+++ src/ScottPlot5/ScottPlot5/Grids/DefaultGrid.cs
@@ -3,26 +3,45 @@
 public class DefaultGrid(IXAxis xAxis, IYAxis yAxis) : IGrid
 {
     public bool IsVisible { get; set; } = true;
-    public LineStyle MajorLineStyle = new() { Width = 1, Color = Colors.Black.WithOpacity(.1) };
-    public LineStyle MinorLineStyle = new() { Width = 0, Color = Colors.Black.WithOpacity(.05) };
-
-    public int MaximumNumberOfGridLines = 1000;
-
     public bool IsBeneathPlottables { get; set; } = true;
-
     public IXAxis XAxis { get; set; } = xAxis;
     public IYAxis YAxis { get; set; } = yAxis;
+    public GridStyle XAxisStyle { get; set; } = new();
+    public GridStyle YAxisStyle { get; set; } = new();
 
-    /// <summary>
-    /// Controls color for <see cref="MajorLineStyle"/> and <see cref="MinorLineStyle"/>
-    /// </summary>
-    public Color LineColor
+    public Color MajorLineColor
     {
-        get => MajorLineStyle.Color;
         set
         {
-            MajorLineStyle.Color = value;
-            MinorLineStyle.Color = value;
+            XAxisStyle.MajorLineStyle.Color = value;
+            YAxisStyle.MajorLineStyle.Color = value;
+        }
+    }
+
+    public Color MinorLineColor
+    {
+        set
+        {
+            XAxisStyle.MinorLineStyle.Color = value;
+            YAxisStyle.MinorLineStyle.Color = value;
+        }
+    }
+
+    public float MajorLineWidth
+    {
+        set
+        {
+            XAxisStyle.MinorLineStyle.Width = value;
+            YAxisStyle.MinorLineStyle.Width = value;
+        }
+    }
+
+    public float MinorLineWidth
+    {
+        set
+        {
+            XAxisStyle.MinorLineStyle.Width = value;
+            YAxisStyle.MinorLineStyle.Width = value;
         }
     }
 
@@ -39,35 +58,7 @@ public class DefaultGrid(IXAxis xAxis, IYAxis yAxis) : IGrid
         var xTicks = XAxis.TickGenerator.Ticks.Where(x => x.Position >= minX && x.Position <= maxX);
         var yTicks = YAxis.TickGenerator.Ticks.Where(x => x.Position >= minY && x.Position <= maxY);
 
-        if (MinorLineStyle.Width > 0)
-        {
-            float[] xTicksMinor = xTicks.Where(x => !x.IsMajor).Select(x => XAxis.GetPixel(x.Position, rp.DataRect)).ToArray();
-            float[] yTicksMinor = yTicks.Where(x => !x.IsMajor).Select(x => YAxis.GetPixel(x.Position, rp.DataRect)).ToArray();
-            RenderGridLines(rp, xTicksMinor, XAxis.Edge, MinorLineStyle);
-            RenderGridLines(rp, yTicksMinor, YAxis.Edge, MinorLineStyle);
-        }
-
-        if (MajorLineStyle.Width > 0)
-        {
-            float[] xTicksMajor = xTicks.Where(x => x.IsMajor).Select(x => XAxis.GetPixel(x.Position, rp.DataRect)).ToArray();
-            float[] yTicksMajor = yTicks.Where(x => x.IsMajor).Select(x => YAxis.GetPixel(x.Position, rp.DataRect)).ToArray();
-            RenderGridLines(rp, xTicksMajor, XAxis.Edge, MajorLineStyle);
-            RenderGridLines(rp, yTicksMajor, YAxis.Edge, MajorLineStyle);
-        }
-    }
-
-    private void RenderGridLines(RenderPack rp, float[] positions, Edge edge, LineStyle lineStyle)
-    {
-        Pixel[] starts = new Pixel[positions.Length];
-        Pixel[] ends = new Pixel[positions.Length];
-
-        for (int i = 0; i < positions.Length; i++)
-        {
-            float px = positions[i];
-            starts[i] = edge.IsHorizontal() ? new Pixel(px, rp.DataRect.Bottom) : new Pixel(rp.DataRect.Left, px);
-            ends[i] = edge.IsHorizontal() ? new Pixel(px, rp.DataRect.Top) : new Pixel(rp.DataRect.Right, px);
-        }
-
-        Drawing.DrawLines(rp.Canvas, starts, ends, lineStyle.Color, lineStyle.Width, antiAlias: true, lineStyle.Pattern);
+        XAxisStyle.Render(rp, XAxis, xTicks);
+        YAxisStyle.Render(rp, YAxis, yTicks);
     }
 }